Do NOT perform 'cleanup' if cleanup is false

+++ src/upload.ts
@@ -23,12 +23,14 @@ export async function upload(
   }
 
   // Remove blobs under the 'name'
-  for await (const blob of containerClient.listBlobsFlat()) {
-    if (!blob.name.startsWith(name + '/')) {
-      continue;
+  if (cleanup) {
+    for await (const blob of containerClient.listBlobsFlat()) {
+      if (!blob.name.startsWith(name + '/')) {
+        continue;
+      }
+      const blockClient = await containerClient.getBlockBlobClient(blob.name);
+      await blockClient.delete();
     }
-    const blockClient = await containerClient.getBlockBlobClient(blob.name);
-    await blockClient.delete();
   }
 
   // Upload the file/directory
